fix(tests): make it possible to run `npm run test:integration:post` against docker db
