+++ env-vars-testing.conf
@@ -8,7 +8,7 @@ WHYD_URL_PREFIX="http://localhost:8080"
 WHYD_PORT="8080"
 MONGODB_DATABASE="openwhyd_test"
 MONGODB_HOST="localhost"
-MONGODB_PORT="27017"
+MONGODB_PORT="27117"
 MONGODB_USER=""
 MONGODB_PASS=""
 SENDGRID_API_KEY=""
