+++ test/fixtures.js
@@ -1,4 +1,19 @@
+//@ts-check
+
+const fs = require('fs');
 const childProcess = require('child_process');
+
+const readFile = (file) => fs.promises.readFile(file, 'utf-8');
+
+exports.loadEnvVars = async (file) => {
+  const envVars = {};
+  (await readFile(file)).split(/[\r\n]+/).forEach((envVar) => {
+    if (!envVar) return;
+    const [key, def] = envVar.split('=');
+    envVars[key] = def.replace(/^"|"$/g, '');
+  });
+  return envVars;
+};
 
 exports.URL_PREFIX = 'http://localhost:8080';
 
@@ -34,9 +49,30 @@ exports.TEST_USER = {
 };
 
 // Call this before each test to prevent side effects between tests
-exports.cleanup = function (done) {
+exports.cleanup = async function () {
   this.timeout(4000);
   console.warn('🧹 Cleaning up test db...');
-  const process = childProcess.fork('test/reset-test-db.js');
-  process.on('close', () => done());
+  const envFile = process.env.START_WITH_ENV_FILE
+    ? await exports.loadEnvVars(process.env.START_WITH_ENV_FILE)
+    : {};
+  const resetDbProcess = childProcess.fork('test/reset-test-db.js', {
+    env: {
+      ...process.env,
+      MONGODB_HOST: envFile.MONGODB_HOST || process.env.MONGODB_HOST,
+      MONGODB_PORT: envFile.MONGODB_PORT || process.env.MONGODB_PORT,
+    },
+  });
+
+  // resetDbProcess.stdout.on('data', (data) => {
+  //   console.log(`stdout: ${data}`);
+  // });
+
+  // resetDbProcess.stderr.on('data', (data) => {
+  //   console.error(`stderr: ${data}`);
+  // });
+
+  // resetDbProcess.on('error', (err) => console.error('cleanup error:', err));
+  // resetDbProcess.on('close', () => done());
+
+  return new Promise((resolve) => resetDbProcess.on('close', () => resolve()));
 };
